Refactor many things for fun and add extra props to JSON output. (#32)
* Re-add the code but to the correct branch this time.

* Minor refactors in FileSink so it doesn't format time if it doesn't need to.

Replace octo::logger::Log::LogLevel with an alias to octo::logger::LogLevel and implement octo::logger::LogLevelUtils for more efficient to/from string actions.

* Added the changes from PR #27, manager locking related changes.

* Added ConsoleJSONSink implementation.

* Re-added Log::level_to_string and Log::string_to_level with the deprecated attribute, this should take care of backward compatability.

* Added ConsoleJSONSink UT.

* Replace ForkSafeMutex 'operator std::mutex&()' with 'std::mutex& operator*()'.

+++ examples/src/basic_logger.cpp
@@ -16,7 +16,7 @@ int main(int argc, char** argv)
 {
     auto config = std::make_shared<octo::logger::ManagerConfig>();
     config->set_option(octo::logger::ManagerConfig::LoggerOption::DEFAULT_CHANNEL_LEVEL,
-                       static_cast<int>(octo::logger::Log::LogLevel::TRACE));
+                       octo::logger::Log::LogLevel::TRACE);
     octo::logger::SinkConfig console_sink("Console", octo::logger::SinkConfig::SinkType::CONSOLE_SINK);
     config->add_sink(console_sink);
     octo::logger::Manager::instance().configure(config);
